[CMA-114] refactor : ShortsPlayerViewModel 컴포넌트화
