+++ app/src/main/java/com/climus/climeet/presentation/ui/main/shorts/ShortsFragment.kt
@@ -3,6 +3,7 @@ package com.climus.climeet.presentation.ui.main.shorts
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import com.climus.climeet.R
@@ -11,17 +12,22 @@ import com.climus.climeet.presentation.base.BaseFragment
 import com.climus.climeet.presentation.ui.main.global.selectsector.BottomSheetState
 import com.climus.climeet.presentation.ui.main.shorts.adapter.ShortsThumbnailAdapter
 import com.climus.climeet.presentation.ui.main.shorts.adapter.UpdatedFollowAdapter
+import com.climus.climeet.presentation.ui.main.shorts.player.ShortsOption
+import com.climus.climeet.presentation.ui.main.shorts.player.ShortsPlayerEvent
+import com.climus.climeet.presentation.ui.main.shorts.player.ShortsPlayerViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class ShortsFragment : BaseFragment<FragmentShortsBinding>(R.layout.fragment_shorts) {
 
-    private val sharedViewModel: ShortsViewModel by activityViewModels()
+    private val sharedViewModel: ShortsPlayerViewModel by activityViewModels()
+    private val viewModel: ShortsViewModel by viewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.vm = sharedViewModel
+        binding.svm = sharedViewModel
+        binding.vm = viewModel
 
         binding.rvShortsThumbnail.adapter = ShortsThumbnailAdapter()
         binding.rvUpdatedFollow.adapter = UpdatedFollowAdapter()
@@ -45,8 +51,19 @@ class ShortsFragment : BaseFragment<FragmentShortsBinding>(R.layout.fragment_sho
         repeatOnStarted {
             sharedViewModel.event.collect {
                 when (it) {
-                    is ShortsEvent.ShowToastMessage -> showToastMessage(it.msg)
-                    is ShortsEvent.NavigateToShortsDetail -> findNavController().toShortsPlayer(it.shortsId, it.position)
+                    is ShortsPlayerEvent.ShowToastMessage -> showToastMessage(it.msg)
+                    is ShortsPlayerEvent.NavigateToShortsPlayer -> findNavController().toShortsPlayer(
+                        it.shortsId,
+                        it.position
+                    )
+
+                }
+            }
+        }
+
+        repeatOnStarted {
+            viewModel.event.collect {
+                when (it) {
                     is ShortsEvent.NavigateToSearchCragBottomSheet -> {
                         BottomSheetState.state = "SHORTS"
                         findNavController().toShortsBottomSheet()
@@ -56,13 +73,14 @@ class ShortsFragment : BaseFragment<FragmentShortsBinding>(R.layout.fragment_sho
         }
     }
 
-    private fun NavController.toShortsBottomSheet(){
+    private fun NavController.toShortsBottomSheet() {
         val action = ShortsFragmentDirections.actionShortsFragmentToShortsBottomSheetFragment()
         navigate(action)
     }
 
-    private fun NavController.toShortsPlayer(shortsId: Long, position: Int){
-        val action = ShortsFragmentDirections.actionShortsFragmentToShortsPlayerFragment(shortsId, position)
+    private fun NavController.toShortsPlayer(shortsId: Long, position: Int) {
+        val action =
+            ShortsFragmentDirections.actionShortsFragmentToShortsPlayerFragment(shortsId, position)
         navigate(action)
     }
 }
