+++ app/src/main/java/com/climus/climeet/presentation/ui/main/shorts/player/ShortsDetailFragment.kt
@@ -105,6 +105,7 @@ class ShortsDetailFragment @Inject constructor(
                 else -> {}
             }
         }
+
     }
 
     override fun onPause() {
@@ -121,6 +122,5 @@ class ShortsDetailFragment @Inject constructor(
         super.onDestroy()
         releasePlayer()
     }
-
 
 }
