+++ app/src/main/java/com/climus/climeet/presentation/ui/main/shorts/player/ShortsPlayerFragment.kt
@@ -1,26 +1,21 @@
 package com.climus.climeet.presentation.ui.main.shorts.player
 
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.navArgs
 import androidx.viewpager2.widget.ViewPager2
 import com.climus.climeet.R
-import com.climus.climeet.app.App
 import com.climus.climeet.databinding.FragmentShortsPlayerBinding
 import com.climus.climeet.presentation.base.BaseFragment
 import com.climus.climeet.presentation.ui.main.MainViewModel
-import com.climus.climeet.presentation.ui.main.shorts.ShortsOption
-import com.climus.climeet.presentation.ui.main.shorts.ShortsViewModel
 import com.climus.climeet.presentation.ui.main.shorts.adapter.ShortsDetailAdapter
-import com.climus.climeet.presentation.util.Constants.TAG
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class ShortsPlayerFragment: BaseFragment<FragmentShortsPlayerBinding>(R.layout.fragment_shorts_player) {
 
-    private val sharedViewModel: ShortsViewModel by activityViewModels()
+    private val sharedViewModel: ShortsPlayerViewModel by activityViewModels()
     private val parentViewModel: MainViewModel by activityViewModels()
     private var adapter : ShortsDetailAdapter? = null
