+++ app/src/main/res/layout/fragment_shorts_detail.xml
@@ -21,6 +21,7 @@
             android:id="@+id/exo_player"
             android:layout_width="0dp"
             android:layout_height="0dp"
+            app:controller_layout_id="@layout/custom_progress_ui"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -121,7 +122,7 @@
             android:layout_height="90dp"
             android:layout_marginEnd="20dp"
             android:layout_marginBottom="20dp"
-            app:layout_constraintBottom_toTopOf="@id/pb_video"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:routeImgUrl="@{item.routeImgUrl}"
             app:routeLevelColor="@{item.gymLevelColor}"
@@ -177,7 +178,7 @@
             android:text="@{item.gymName}"
             android:visibility="@{item.gymName.empty ? View.INVISIBLE : View.VISIBLE}"
             app:drawableStartCompat="@drawable/ic_location_small"
-            app:layout_constraintBottom_toTopOf="@id/pb_video"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             tools:text="더클라임 연남" />
 
@@ -196,22 +197,22 @@
             android:text="@{item.sectorName}"
             android:visibility="@{item.sectorName.empty ? View.INVISIBLE : View.VISIBLE}"
             app:drawableStartCompat="@drawable/ic_walking"
-            app:layout_constraintBottom_toTopOf="@id/pb_video"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toEndOf="@id/tv_gym_name"
             tools:text="SECTOR 2-2" />
 
-        <ProgressBar
-            android:id="@+id/pb_video"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:max="100"
-            android:progressBackgroundTint="@color/white"
-            android:progressTint="@color/cm_main"
-            android:scaleY="1"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
+<!--        <ProgressBar-->
+<!--            android:id="@+id/pb_video"-->
+<!--            style="?android:attr/progressBarStyleHorizontal"-->
+<!--            android:layout_width="0dp"-->
+<!--            android:layout_height="wrap_content"-->
+<!--            android:max="100"-->
+<!--            android:progressBackgroundTint="@color/white"-->
+<!--            android:progressTint="@color/cm_main"-->
+<!--            android:scaleY="1"-->
+<!--            app:layout_constraintBottom_toBottomOf="parent"-->
+<!--            app:layout_constraintEnd_toEndOf="parent"-->
+<!--            app:layout_constraintStart_toStartOf="parent" />-->
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
