+++ app/src/main/res/layout/fragment_shorts_player.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="vm"
-            type="com.climus.climeet.presentation.ui.main.shorts.ShortsViewModel" />
+            type="com.climus.climeet.presentation.ui.main.shorts.player.ShortsPlayerViewModel" />
 
     </data>
